refactor: update Prisma import paths and initialize PrismaClient in constants
