+++ apps/server/src/app.ts
@@ -1,16 +1,13 @@
 import { handle } from 'i18next-http-middleware';
 import { serve, setup } from 'swagger-ui-express';
 
-import { PrismaClient } from '@prisma/client';
 import { authApi } from '@server/api';
 import { zodMiddleware } from '@server/api/middlewares';
 import { i18n } from '@server/i18n/i18n';
 import { authRouter } from '@server/routes';
-import { getPortOfServer } from '@server/utils';
+import { getPortOfServer, prisma } from '@server/utils';
 import { zodiosApp } from '@zodios/express';
 import { bearerAuthScheme, openApiBuilder } from '@zodios/openapi';
-
-export const prisma = new PrismaClient();
 
 const app = zodiosApp();
