+++ apps/server/src/utils/constants.ts
@@ -1,36 +1,13 @@
 import { configDotenv } from 'dotenv';
 import path from 'path';
-import pkg from 'pg';
 import { fileURLToPath } from 'url';
+
+import { PrismaClient } from '@prisma/client';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-const { Pool } = pkg;
 
-// Load root .env first, then local .env (local overrides root)
-configDotenv({ path: path.resolve(__dirname, '../../../../.env') });
 configDotenv({ path: path.resolve(__dirname, '../../.env') });
-
-const getPostgresConfig = () => {
-  const user = process.env.POSTGRES_USER;
-  const host = process.env.POSTGRES_HOST;
-  const database = process.env.POSTGRES_DB;
-  const password = process.env.POSTGRES_PASSWORD;
-  const port = Number(process.env.POSTGRES_PORT);
-
-  if (!user || !host || !database || !password || isNaN(port) || port <= 0) {
-    console.error('Invalid PostgreSQL configuration in environment variables.');
-    process.exit(1);
-  }
-
-  return {
-    user,
-    host,
-    database,
-    password,
-    port
-  };
-};
 
 export const getSaltRounds = () => {
   const saltRounds = Number(process.env.SALT_ROUNDS);
@@ -54,7 +31,7 @@ export const getPortOfServer = () => {
   return port;
 };
 
-export const pool = new Pool(getPostgresConfig());
+export const prisma = new PrismaClient();
 
 export const LOCKOUT_TIERS_MINUTES = [5, 15, 60, 180, 720, 1440]; // up to 24 hours
 export const MAX_TIERS = LOCKOUT_TIERS_MINUTES.length;
